Decoder 마다 Greedy/BeamSearch 구현

+++ models/clean-ASR/modules/decoder/label_smoothing_loss.py
@@ -56,7 +56,7 @@ class LabelSmoothingLoss(nn.Module):
             true_dist.fill_(self.smoothing / (self.size - 1))
             ignore = target == self.padding_idx  # (B,)
             total = len(target) - ignore.sum().item()
-            target = target.masked_fill(ignore, 0)  # avoid -1 index
+            target = target.masked_fill(ignore, 0).to(torch.long)  # avoid -1 index
             true_dist.scatter_(1, target.unsqueeze(1), self.confidence)
         kl = self.criterion(torch.log_softmax(x, dim=1), true_dist)
         denom = total if self.normalize_length else batch_size
